Longer commit message this time, because it's a big change and I want to talk about the project's situation. When I first completed challenge #12, I accidentally used secret data that should have been inaccessible. I rewrote it from (mostly) scratch, with more segmentation into functions and a more clean and mature coding style after reading some glibc code and writing proper python pwnlib exploit code for exploit.education's Protostar machine.
I needed to come back to this challenge with a fresh perspective in order to
approach it from a different direction, but did not want to move on without
completing it properly, particularly because challenge #14 builds on #12. Now
that it has been resolved, I plan to end my hiatus on this project and complete
the second set of problems.

Summer internship will be keeping me busy, but my goal is to complete #16 before
the beginning of the Fall term in August.