Removed stdint.h, which does not conform to C89. My compiler, gcc, did not complain because GNU's implementation of C89, gnu89, includes stdint.h and thus does not require code to conform strictly to the C89 standard.
crypto_utility.h now includes directives to attempt to typedef unsigned char to
uint8_t and either int or long to uint32_t based on maximums from limits.h. It's
kind of a hack job, but it should work reasonably well on most machines and was
all I was using from that header.

All programs have been recompiled. Some were modified slightly to work with
updated headers. The repository is now ready for work to begin on Set 3.